Second day done

+++ src/com/Jonas/SJGE/tilemap/tiles/Tile.java
@@ -22,6 +22,6 @@ public abstract class Tile {
 		renderTile(screen, tilemapLocation, x, y);
 	}
 	protected void renderTile(Screen screen, int tilemapLocation, int x, int y) {
-		screen.screen.draw(ImageLoader.tilemap, x * Tilemap.TILE_SIZE - (int) game.cam.x, y * Tilemap.TILE_SIZE - (int) game.cam.y, (tilemapLocation % 16) * Tilemap.TILE_SIZE, (tilemapLocation / 16) * Tilemap.TILE_SIZE, Tilemap.TILE_SIZE, Tilemap.TILE_SIZE);
+		screen.screen.draw(ImageLoader.tilemap, x * Tilemap.TILE_SIZE - game.cam.x, y * Tilemap.TILE_SIZE - game.cam.y, (tilemapLocation % 16) * Tilemap.TILE_SIZE, (tilemapLocation / 16) * Tilemap.TILE_SIZE, Tilemap.TILE_SIZE, Tilemap.TILE_SIZE);
 	}
 }
